Version 0.3.0, with a lot of improvements
For users:
There are now two different Color images: the first is generated
as an attempt, the second is the result of merging pixels which switch
in consecutive frames. These can be computed automatically using
self.gostbusters() (see __doc__ for details),
or manually with
self.manualGhostbusters, which asks for the image number to join
with the next one.
The switch time is now available directly from the image
on the bottom message.

Added the module self.showRawAndCalcImages
to directly compare raw and calculated images


The distributions of avalanches and clusters
are now correctly calculated.

For the developer:
Added a lot of 'internal' functions using self._something
* self._switchTimes is the 1D array of the switch times
* self._switchIimes2D is the 2D image of the switch times,
considering the threshold and
* self.imageDir is calculated correctly
* getAxyLabels revisited and simplified